refactor(dashboard): split into package, redesign inspector, fix sim
Split the 2849-line dashboard.py into a proper package with frontend
assets as real files:

  src/openrange/dashboard/
    events.py       DashboardEvent, EventBridge, DashboardArtifactLog,
                    read/write helpers, fallback_narrate
    topology.py     normalized_runtime_topology, public_world,
                    normalized_rows/strings, embedded_topology
    summaries.py    activity_summary, actor_summaries, health_summary
    view.py         DashboardView
    server.py       HTTP server + handler with static file routing
    static/index.html, dashboard.css, dashboard.js

Frontend files now have proper editor/lint/format support and
hot-reload (server reads them on each request). Public import surface
is unchanged.

Inspector ("under the sim") redesign:

  Before: topbar with duplicate title/Reset/Pause/Play, 5-column
  status strip, 9 cards in a grid each containing a small summary
  plus a giant <pre> JSON dump of the same data, plus a tabbed
  Details panel with raw JSON. ~70% of every card was unreadable
  indented JSON. Activity / Actors / Narration cards duplicated
  the sim's overlays.

  After: thin Snapshot identity bar (id + counts + admission pill),
  Briefing hero (title/goal/entrypoint chips), Tasks / Admission
  side-by-side, World (services grouped by zone) / Lineage (timeline)
  side-by-side, Artifacts grid. Every <pre> JSON dump removed.
  Restyled to match the sim's dark/blocky aesthetic so the page
  reads as one app. <main hidden> when no snapshot is loaded
  (instead of 8 dashed empty boxes).

Sim fixes:

  - stationDefinitions: drop the loop that pulled the first 4
    artifact paths into the world graph as desks (README.md and
    kind-config.yaml have no business being network hosts).
  - animateSimulation: NPCs now idle-wander every 4-9s when they
    have no event-driven target. Agents and system actors stay
    deterministic. Purely visual; no events spawned.
  - Clamp dt to 0.1s to prevent characters teleporting off the
    floor when requestAnimationFrame resumes after a tab pause.

New static-file route test asserts the dashboard serves /,
/static/dashboard.{css,js}, and returns 404 for missing assets and
path traversal attempts.

.gitignore: add .claude/ (local-only worktree config).
