+++ .gitignore
@@ -4,6 +4,7 @@
 # Editors
 .idea/
 .vscode/
+.claude/
 *.swp
 *.swo
