+++ src/openrange/dashboard/__init__.py
@@ -1,0 +1,46 @@
+"""Dashboard inspection helpers for admitted snapshots."""
+
+from openrange.dashboard.events import (
+    DashboardArtifactLog,
+    DashboardEvent,
+    EventBridge,
+    dashboard_event_from_mapping,
+    fallback_narrate,
+    read_dashboard_events,
+    read_dashboard_state,
+    write_dashboard_state,
+)
+from openrange.dashboard.server import DashboardHTTPServer, DashboardRequestHandler
+from openrange.dashboard.summaries import (
+    activity_summary,
+    actor_summaries,
+    health_summary,
+    percent_value,
+)
+from openrange.dashboard.topology import (
+    normalized_rows,
+    normalized_strings,
+    public_world,
+)
+from openrange.dashboard.view import DashboardView
+
+__all__ = [
+    "DashboardArtifactLog",
+    "DashboardEvent",
+    "DashboardHTTPServer",
+    "DashboardRequestHandler",
+    "DashboardView",
+    "EventBridge",
+    "activity_summary",
+    "actor_summaries",
+    "dashboard_event_from_mapping",
+    "fallback_narrate",
+    "health_summary",
+    "normalized_rows",
+    "normalized_strings",
+    "percent_value",
+    "public_world",
+    "read_dashboard_events",
+    "read_dashboard_state",
+    "write_dashboard_state",
+]
