+++ src/openrange/dashboard/static/index.html
@@ -1,0 +1,141 @@
+<!doctype html>
+<html lang="en">
+<head>
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>OpenRange Dashboard</title>
+<link rel="preconnect" href="https://fonts.googleapis.com">
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+<link
+  href="https://fonts.googleapis.com/css2?family=Nunito:wght@500;700;800;900&display=swap"
+  rel="stylesheet">
+<script
+  src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.min.js"></script>
+<script
+  src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/controls/OrbitControls.js"></script>
+<link rel="stylesheet" href="/static/dashboard.css">
+</head>
+<body>
+<div class="sim-stage" id="sim-stage">
+  <canvas id="sim-canvas" aria-label="OpenRange episode simulation"></canvas>
+
+  <div class="sim-overlay sim-panel sim-title">
+    <h1>OpenRange Episode Dashboard</h1>
+    <p id="sim-subtitle">Loading public episode briefing</p>
+  </div>
+
+  <div class="sim-overlay sim-panel sim-clock">
+    <strong id="sim-clock">00:00</strong>
+    <span id="sim-status">Waiting</span>
+  </div>
+
+  <div class="sim-overlay sim-panel sim-event-panel">
+    <div class="sim-panel-title">Live Event Feed</div>
+    <ul class="sim-event-list" id="sim-event-log"></ul>
+  </div>
+
+  <div class="sim-overlay sim-panel sim-actor-panel" id="sim-actor-panel">
+    <div class="sim-actor-head">
+      <div>
+        <h2 id="sim-actor-name">Actor</h2>
+        <p id="sim-actor-role">No actor selected</p>
+      </div>
+      <button class="sim-close" id="sim-actor-close" title="Close">X</button>
+    </div>
+    <dl class="sim-actor-grid">
+      <dt>Kind</dt><dd id="sim-actor-kind">event</dd>
+      <dt>Events</dt><dd id="sim-actor-events">0</dd>
+      <dt>Targets</dt><dd id="sim-actor-targets">none</dd>
+      <dt>Latest</dt><dd id="sim-actor-latest">No activity yet.</dd>
+    </dl>
+    <div class="sim-panel-title">Recent History</div>
+    <ul class="sim-actor-history" id="sim-actor-history"></ul>
+  </div>
+
+  <div class="sim-overlay sim-panel sim-empty" id="sim-empty">
+    No admitted world is loaded yet. Start an eval dashboard through the environment
+    or launch against a snapshot store.
+  </div>
+
+  <div class="sim-overlay sim-bottom">
+    <div class="sim-panel sim-narrator">
+      <div class="sim-panel-title">Episode Narrator</div>
+      <p class="sim-narrator-text" id="sim-narrator">No episode activity yet.</p>
+    </div>
+
+    <div class="sim-controls" aria-label="Episode controls">
+      <button class="danger" data-action="reset" title="Reset episode">RST</button>
+      <button data-action="pause" title="Pause episode">PAU</button>
+      <button class="primary" data-action="play" title="Play episode">PLY</button>
+    </div>
+
+    <div class="sim-panel sim-gauges">
+      <div class="sim-panel-title">Run Health</div>
+      <div class="sim-gauge"><span>Uptime</span><i id="sim-uptime-gauge"></i></div>
+      <div class="sim-gauge"><span>Defense</span><i id="sim-defense-gauge"></i></div>
+      <div class="sim-gauge">
+        <span>Integrity</span><i id="sim-integrity-gauge"></i>
+      </div>
+    </div>
+  </div>
+</div>
+
+<main id="inspector" hidden>
+<div class="shell">
+  <header class="identity">
+    <div class="identity-main">
+      <span class="identity-label">Snapshot</span>
+      <code class="identity-id" id="snapshot-id">—</code>
+    </div>
+    <div class="identity-stats" id="identity-stats"></div>
+    <span class="pill" id="admission-pill">Waiting</span>
+  </header>
+
+  <section class="briefing-hero" id="briefing-hero"></section>
+
+  <div class="grid">
+    <section class="card">
+      <header class="card-head">
+        <h2>Tasks</h2>
+        <span class="count" id="task-count">0</span>
+      </header>
+      <div class="card-body" id="tasks"></div>
+    </section>
+
+    <section class="card">
+      <header class="card-head">
+        <h2>Admission</h2>
+        <span class="count" id="admission-count">0 / 0</span>
+      </header>
+      <div class="card-body" id="admission"></div>
+    </section>
+
+    <section class="card">
+      <header class="card-head">
+        <h2>World</h2>
+        <span class="count" id="world-count">0 services</span>
+      </header>
+      <div class="card-body" id="world"></div>
+    </section>
+
+    <section class="card">
+      <header class="card-head">
+        <h2>Lineage</h2>
+        <span class="count" id="lineage-count">0</span>
+      </header>
+      <div class="card-body" id="lineage"></div>
+    </section>
+  </div>
+
+  <section class="card card-wide">
+    <header class="card-head">
+      <h2>Artifacts</h2>
+      <span class="count" id="artifact-count">0</span>
+    </header>
+    <div class="card-body" id="artifacts"></div>
+  </section>
+</div>
+</main>
+<script src="/static/dashboard.js"></script>
+</body>
+</html>
